Move workspace pid code to root.c
