+++ include/sway/tree/workspace.h
@@ -44,10 +44,6 @@ void workspace_output_add_priority(struct sway_container *workspace,
 struct sway_container *workspace_output_get_highest_available(
 		struct sway_container *ws, struct sway_container *exclude);
 
-struct sway_container *workspace_for_pid(pid_t pid);
-
-void workspace_record_pid(pid_t pid);
-
 void workspace_detect_urgent(struct sway_container *workspace);
 
 #endif
